+++ tests/doc_test/doc_needextend/conf.py
@@ -19,6 +19,8 @@
 import os
 import sys
 
+from docutils.parsers.rst import directives
+
 sys.path.insert(0, os.path.abspath("../../sphinxcontrib"))
 
 # -- General configuration ------------------------------------------------
@@ -41,7 +43,9 @@ needs_types = [
     dict(directive="test", title="Test Case", prefix="TC_", color="#DCB239", style="node"),
 ]
 
-needs_extra_links = [{"option": "checks", "incoming": "is checked by", "outgoing": "checks"}]
+needs_extra_options = {
+    "author": directives.unchanged,
+}
 
 plantuml = "java -jar %s" % os.path.join(os.path.dirname(__file__), "..", "..", "..", "docs", "utils", "plantuml.jar")
 plantuml_output_format = "svg"